bug fixing

+++ app/src/main/java/com/example/android/sunshine/utilities/SunshineWeatherUtils.java
@@ -103,7 +103,7 @@ public final class SunshineWeatherUtils {
      *
      * @return Wind String in the following form: "2 km/h SW"
      */
-    public static String getFormattedWind(Context context, float windSpeed, float degrees) {
+    public static String getFormattedWind(Context context, double windSpeed, double degrees) {
         int windFormat = R.string.format_wind_kmh;
 
         if (!SunshinePreferences.isMetric(context)) {
